+++ src/maicos/__init__.py
@@ -9,15 +9,11 @@
 
 import warnings
 
-from ._version import get_versions
+from ._version import __version__  # noqa: F401
 from .modules import *  # noqa: F403
 from .modules import __all__ as __all__
 
 __authors__ = "MAICoS Developer Team"
-#: Version information for MAICoS, following :pep:`440`
-#: and `semantic versioning <http://semver.org/>`_.
-__version__ = get_versions()["version"]
-del get_versions
 
 # Print maicos DeprecationWarnings
 warnings.filterwarnings(action="once", category=DeprecationWarning, module="maicos")